+++ R/add_cost_penalties.R
@@ -127,8 +127,8 @@ NULL
 #'   mp %>%
 #'   add_hier_approach(rel_tol = rel_tol_matrix)
 #'
-#' # generate solutions
-#' ms <- solve(mp)
+#' # generate solutions and remove duplicates
+#' ms <- solve(mp, remove_duplicates = TRUE)
 #'
 #' # plot the solutions
 #' plot(terra::rast(ms), axes = FALSE)
@@ -136,8 +136,8 @@ NULL
 #' # extract objective values for the solutions
 #' obj_matrix <- attributes(ms)$objective
 #'
-#' # preview the objective values
-#' head(obj_matrix)
+#' # print the objective values
+#' print(obj_matrix)
 #'
 #' # plot the objectives values to visualize trade-offs
 #' # (note that smaller values are better for both objectives)