+++ R/add_hier_approach.R
@@ -320,14 +320,13 @@ NULL
 #' # we will explore trade-offs between the two objectives, by generating
 #' # multiple solutions using multi-objective optimization
 #'
-#' # create a matrix with 40 different combinations of relative tolerance values
-#' # that can be used to generate 40 solutions
+#' # create a matrix with multiple different relative tolerance values
 #' rel_tol_matrix <- approach_rel_tol_matrix(
-#'   n_problems = 2, n_values = 40, max = 1.2
+#'   n_problems = 2, n_values = 20, max = 1.2
 #' )
 #'
-#' # preview matrix with relative tolerance values
-#' head(rel_tol_matrix)
+#' # print matrix with relative tolerance values
+#' print(rel_tol_matrix)
 #'
 #' # create a multi-objective problem with the matrix of relative tolerance
 #' # values and - because we do not specify values for priority - the
@@ -335,14 +334,11 @@ NULL
 #' # specified in order of priority
 #' mp2 <-
 #'   multi_problem(keystone_obj = p1, iconic_obj = p2) %>%
-#'   add_hier_approach(
-#'     rel_tol = rel_tol_matrix,
-#'     verbose = FALSE
-#'   ) %>%
-#'   add_default_solver(verbose = FALSE)
+#'   add_hier_approach(rel_tol = rel_tol_matrix, verbose = TRUE) %>%
+#'   add_default_solver(gap = 0.01, verbose = FALSE)
 #'
-#' # solve multi-objective problem and generate 40 solutions
-#' ms2 <- solve(mp2)
+#' # solve multi-objective problem and remove duplicate solutions
+#' ms2 <- solve(mp2, remove_duplicates = TRUE)
 #'
 #' # plot multiple solutions
 #' plot(terra::rast(ms2), axes = FALSE)
@@ -350,8 +346,8 @@ NULL
 #' # extract objective values for the solutions
 #' obj_matrix <- attributes(ms2)$objective
 #'
-#' # preview the objective values
-#' head(obj_matrix)
+#' # print the objective values
+#' print(obj_matrix)
 #'
 #' # plot the objectives values to visualize trade-offs
 #' # (note that smaller values are better because these objectives seek to
@@ -494,6 +490,7 @@ add_hier_approach <- function(x, rel_tol, priority = NULL, verbose = TRUE) {
           ## if needed, set up progress bar
           if (isTRUE(verbose)) {
             pb <- cli::cli_progress_bar(
+              format = cli_progress_bar_format("Generating solutions"),
               total = nrow(rel_tol),
               .envir = parent.frame()
             )