+++ R/add_ref_point_approach.R
@@ -190,18 +190,18 @@ NULL
 #'   n_problems = 2, n_values = 5, include_zero = TRUE
 #' )
 #'
-#' # preview weight matrix
-#' head(weights_matrix)
+#' # print weight matrix
+#' print(weights_matrix)
 #'
 #' # now create multi-objective problem with reference point approach,
 #' # with weights to generate multiple solutions
 #' mp2 <-
 #'   multi_problem(keystone_obj = p1, iconic_obj = p2) %>%
 #'   add_ref_point_approach(weights = weights_matrix, verbose = TRUE) %>%
-#'   add_default_solver(verbose = FALSE)
+#'   add_default_solver(gap = 0.01, verbose = FALSE)
 #'
-#' # solve problem
-#' ms2 <- solve(mp2)
+#' # solve multi-objective problem and remove duplicate solutions
+#' ms2 <- solve(mp2, remove_duplicates = TRUE)
 #'
 #' # plot multiple solutions
 #' plot(terra::rast(ms2), axes = FALSE)
@@ -209,8 +209,8 @@ NULL
 #' # extract objective values for the solutions
 #' obj_matrix <- attributes(ms2)$objective
 #'
-#' # preview the objective values
-#' head(obj_matrix)
+#' # print the objective values
+#' print(obj_matrix)
 #'
 #' # plot the objectives values to visualize trade-offs
 #' # (note that smaller values are better because these objectives seek to
@@ -427,7 +427,9 @@ add_ref_point_approach <- function(x,
           ## if needed, set up progress bar
           if (isTRUE(verbose)) {
             pb <- cli::cli_progress_bar(
-              "Generating solutions", total = nrow(weights)
+              format = cli_progress_bar_format("Generating solutions"),
+              total = nrow(weights),
+              .envir = parent.frame()
             )
           }
           ## calculate shortfall bounds