+++ R/add_wtd_sum_approach.R
@@ -178,17 +178,17 @@ NULL
 #'   n_problems = 2, n_values = 5, include_zero = TRUE
 #' )
 #'
-#' # preview weight matrix
-#' head(weights_matrix)
+#' # print weight matrix
+#' print(weights_matrix)
 #'
 #' # create multi-objective problem using weight matrix
 #' mp2 <-
 #'   multi_problem(keystone_obj = p1, iconic_obj = p2) %>%
-#'   add_wtd_sum_approach(weights_matrix, verbose = FALSE) %>%
-#'   add_default_solver(verbose = FALSE)
+#'   add_wtd_sum_approach(weights_matrix, verbose = TRUE) %>%
+#'   add_default_solver(gap = 0.01, verbose = FALSE)
 #'
-#' # solve multi-objective problem and generate multiple solutions
-#' ms2 <- solve(mp2)
+#' # solve multi-objective problem and remove duplicate solutions
+#' ms2 <- solve(mp2, remove_duplicates = TRUE)
 #'
 #' # plot multiple solutions
 #' plot(terra::rast(ms2), axes = FALSE)
@@ -196,8 +196,8 @@ NULL
 #' # extract objective values for the solutions
 #' obj_matrix <- attributes(ms2)$objective
 #'
-#' # preview the objective values
-#' head(obj_matrix)
+#' # print the objective values
+#' print(obj_matrix)
 #'
 #' # plot the objectives values to visualize trade-offs
 #' # (note that smaller values are better because these objectives seek to
@@ -285,6 +285,7 @@ add_wtd_sum_approach <- function(x, weights, verbose = TRUE) {
           ## if needed, set up progress bar
           if (isTRUE(verbose)) {
             pb <- cli::cli_progress_bar(
+              format = cli_progress_bar_format("Generating solutions"),
               total = nrow(weights),
               .envir = parent.frame()
             )