+++ R/multi_problem.R
@@ -112,13 +112,13 @@ NULL
 #'       add_binary_decisions()
 #'   ) %>%
 #'   add_hier_approach(rel_tol = 0.1, verbose = FALSE) %>%
-#'   add_gurobi_solver(gap = 0, verbose = FALSE)
+#'   add_default_solver(gap = 0, verbose = FALSE)
 #'
 #' # solve problem
 #' ms <- solve(mp)
 #'
 #' # plot solution
-#' plot(ms, axes = FALSE)
+#' plot(ms, main = "solution", axes = FALSE)
 #'
 #' @export
 multi_problem <- function(..., problem_names = NULL) {